implement poor men's aspects
This checkin adds new blocks of the form
  "begin_aspect aspect_name exp else_aspect exp_else end"
and
  "begin_aspect aspect_name exp end"
The idea is to implement very simple aspects. By providing
a command-line argument "-incl_aspect aspect_name" or
"-excl_aspect aspect_name", one can configure, whether
these aspects blocks are used. By default all aspects
are included.

If "aspect_name" is included,
  "begin_aspect aspect_name exp else_aspect exp_else end"
becomes "exp" otherwise "exp_else". For
  "begin_aspect aspect_name exp end"
it becomes "exp" and is otherwise deleted. This deleting
is problematic. It is only possible, if it occurs in a
position, whether this does not change the type of the
containing expression. It is supported only, if the block
appears in explicitly enumerated set or list or as an
argument to a binary operator, such that the other argument
has the same type as the whole expression (e.g. boolean
conjunction and disjunction) and is not an aspect
block itself.